add event case class #3

+++ conf/evolutions/default/1.sql
@@ -2,7 +2,7 @@
 # --- !Ups
 
 CREATE TABLE "user" (
-  "id" INT(20) AUTO_INCREMENT NOT NULL,
+  "id" BIGINT AUTO_INCREMENT NOT NULL,
   "name" VARCHAR(255) NOT NULL,
   "email" VARCHAR(255) NOT NULL,
   "age" INT NOT NULL,
@@ -11,6 +11,8 @@ CREATE TABLE "user" (
 
 INSERT INTO "user" VALUES (NULL, 'Lambda', 'lambda@gmail.com', 27);
 INSERT INTO "user" VALUES (NULL, 'Noh', 'noh@gmail.com', 20);
+
+
 
 # --- !Downs
